templates/advanced/requirements.txt
pyyaml==6.0.1
tqdm==4.67.1
